src/requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.8
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
Brotli==1.0.9
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
click==7.1.2
cryptography==3.4.7
decorator==5.0.6
discord.py==1.6.0
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Compress==1.9.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
idna==3.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.functools==3.3.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
ldap3==2.9
Mako==1.1.4
MarkupSafe==1.1.1
more-itertools==8.7.0
multidict==5.1.0
numpy==1.20.2
pandas==1.2.3
paramiko==2.7.2
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pgadmin4==5.1
pickleshare==0.7.5
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.8.1
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
pyzmq==22.0.3
simplejson==3.17.2
six==1.15.0
speaklater==1.3
speaklater3==1.4
SQLAlchemy==1.3.24
sqlparse==0.4.1
sshtunnel==0.4.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.3
